refactor: add directory annotations and move LogExecutTime
